+++ Preferences.py
@@ -11,6 +11,7 @@ DEFAULTS = {
     'randomize': True,  # Randomize screenshot names?
     'copyonly': False,  # Copy (don't move) screen shots.
     'launchAtStartup': True,
+    'iconset': 'default',  # Which status bar icon? 'default' or 'grayscale'
 }
 
 
@@ -22,6 +23,7 @@ class PreferencesWindowController(UpShotWindowController):
     dropboxdir = objc.IBOutlet()
     dropboxid = objc.IBOutlet()
     launchAtStartup = objc.IBOutlet()
+    iconset = objc.IBOutlet()
 
     def showWindow_(self, sender):
         super(PreferencesWindowController, self).showWindow_(sender)
@@ -29,9 +31,12 @@ class PreferencesWindowController(UpShotWindowController):
 
     def updateDisplay(self):
         """Update window display from settings."""
+        self.launchAtStartup.setState_(get_pref('launchAtStartup'))
+        self.iconset.selectCellWithTag_(
+            1 if get_pref('iconset') == 'grayscale' else 0)
+
         self.randomize.setState_(get_pref('randomize'))
         self.copyonly.setState_(get_pref('copyonly'))
-        self.launchAtStartup.setState_(get_pref('launchAtStartup'))
 
         dropboxdir = detect_dropbox_folder()
         self.dropboxdir.setStringValue_(
@@ -41,11 +46,20 @@ class PreferencesWindowController(UpShotWindowController):
     @objc.IBAction
     def saveSettings_(self, sender):
         """Save changed settings."""
-        set_pref('randomize', bool(self.randomize.state()))
-        set_pref('copyonly', bool(self.copyonly.state()))
-
         set_pref('launchAtStartup', bool(self.launchAtStartup.state()))
         launch_at_startup(bool(self.launchAtStartup.state()))
+
+        # Iconset
+        iconset_sel = self.iconset.selectedCell().tag()
+        if iconset_sel == 1:  # Grayscale
+            set_pref('iconset', 'grayscale')
+        else:
+            set_pref('iconset', 'default')
+        upshot = NSApplication.sharedApplication().delegate()
+        upshot.update_menu()
+
+        set_pref('randomize', bool(self.randomize.state()))
+        set_pref('copyonly', bool(self.copyonly.state()))
 
         try:
             set_pref('dropboxid', int(self.dropboxid.stringValue()))